Restructured entire game again
