+++ src/Actors/ComaActor.py
@@ -7,4 +7,4 @@ class ComaAI(AbstractActor):
         AbstractActor.__init__(self, cards, max_health)
 
     def take_turn(self):  # noqa
-        print("Coma AI is in a coma and does nothing.")
+        print("Coma AI is at", self.current_health, "in a coma and does nothing.")
