+++ src/Monsters/AbstractMonster.py
