+++ src/Monsters/__init__.py
@@ -1,0 +1,4 @@
+from src.Monsters.AbstractMonster import *
+from src.Monsters.JawWorm import *
+from src.Monsters.Sentry import *
+from src.Monsters.TestMonster import *
